bug:638,619,607,564,563,554,544

+++ src/views/people-management/people-set.vue
@@ -685,16 +685,28 @@
 					this.$message.error("只能选择一个删除");
 					return;
 				}
-				this.$api.jobUserManagement.deletePeopleMsg(this.multipleSelection[0].toid).then(res => {
-						if((res.data.data.msg = "success")) {
-							this.$message.success("删除成功");
-							//刷新表格
-							this.getTableData("");
-						}
-					}),
-					error => {
-						console.log(error);
-					};
+                this.$confirm('此操作将永久删除该人员, 是否继续?', '提示', {
+                    confirmButtonText: '确定',
+                    cancelButtonText: '取消',
+                    type: 'warning'
+                }).then(() => {
+                    this.$api.jobUserManagement.deletePeopleMsg(this.multipleSelection[0].toid).then(res => {
+                        if((res.data.data.msg = "success")) {
+                            this.$message.success("删除成功");
+                            //刷新表格
+                            this.getTableData("");
+                        }
+                    }),
+                    error => {
+                        console.log(error);
+                    };
+                }).catch(() => {
+                    this.$message({
+                        type: 'info',
+                        message: '已取消删除'
+                    });
+                });
+
 			},
 
 			//编辑
